Split the dataset into training and testing sets based on a strata variable

Splits the data into testing/training and ensures variables are appropriate for classification algorithm

split_dataset(data_set, strata_variable, predictor)

Arguments

data_set

the data set to be used for the algorithm

strata_variable

a string representing the variable that will be set as the strata in initial_split

predictor

a string representing the variable to be used as the predictor for the algorithm

Value

a list containing the training and the testing data sets

Examples

data_set <- data.frame(
x = rnorm(100),
y = sample(c("A", "B", "C"), 100, replace = TRUE),
z = rnorm(100)
)
strata_variable <- "y"
predictor <- "x"

split_dataset(data_set, strata_variable, predictor)
#> [[1]]
#>               x y            z
#> 1   -0.83023082 A -2.684515346
#> 3   -0.33683102 A -0.406194490
#> 7   -0.24051777 A  0.382802931
#> 10   0.84585411 A -1.232066086
#> 21  -0.37714497 A -1.167589180
#> 22   0.99570791 A  0.877909663
#> 28  -1.51019281 A  0.379080554
#> 29  -2.19225720 A -0.005567858
#> 36  -1.75343241 A  0.281689591
#> 40   1.21991492 A -0.826724356
#> 42  -0.90758815 A  0.589355208
#> 43  -0.44233979 A  0.106140406
#> 44  -0.29828621 A -0.555630389
#> 45  -0.06674028 A -2.525203434
#> 46  -1.24139437 A -0.478187192
#> 47  -0.75781760 A -0.439540923
#> 50  -0.23900527 A -1.085456849
#> 58  -0.05382399 A  0.839180186
#> 60   1.88622143 A -0.067185137
#> 61  -0.76613131 A -0.249501784
#> 68  -1.86534267 A -0.013960490
#> 69   0.05693208 A  0.250868769
#> 78   0.46106369 A  1.147712907
#> 88  -0.45985570 A  0.226036538
#> 91   1.92466577 A  0.502239116
#> 92   1.50823864 A -0.285409195
#> 96  -0.02343607 A  1.266776495
#> 98   0.71844051 A  0.655280412
#> 2   -0.06415357 B -0.495181853
#> 8    2.99119352 B  1.198870927
#> 11  -0.66037070 B  0.063737110
#> 14  -0.31985835 B -2.188785808
#> 16  -0.14669510 B  0.325685635
#> 19   0.84179562 B -1.354017577
#> 20  -0.76811125 B -1.287783377
#> 23   0.90152955 B -0.412485055
#> 25   0.45465683 B -0.525154292
#> 27  -0.47490388 B -0.292009378
#> 37  -0.71782214 B  1.105466051
#> 49  -0.66671218 B  0.644692281
#> 54   0.28666638 B  2.411806471
#> 64  -0.17919090 B  0.590859516
#> 65   0.26530220 B -0.053894866
#> 72  -0.35792732 B  0.184368942
#> 73   1.19056301 B  0.826937412
#> 76  -0.55255443 B  2.111793609
#> 84  -0.04936066 B  2.369412033
#> 86  -0.20259266 B  0.725330842
#> 99  -0.10345661 B -0.367588736
#> 4    0.80796513 C -0.205234143
#> 5    0.71955333 C  1.065972778
#> 9    0.69246612 C -0.927734945
#> 12   1.93030103 C  0.398747282
#> 13  -1.05323049 C  0.927183278
#> 17   0.63661639 C  0.476266134
#> 18   0.79660500 C  0.462674173
#> 26   0.20376264 C  0.151931890
#> 30  -0.33414949 C  1.476280900
#> 35  -0.96481572 C  0.355163821
#> 41  -0.18180305 C  0.398982478
#> 55   0.08896579 C  0.047857345
#> 56   0.18181040 C  2.068948197
#> 62   0.40443706 C  1.100019036
#> 63   1.08391416 C  1.896951298
#> 66  -0.87972646 C -0.164477779
#> 74   0.19679974 C -1.516803337
#> 77   2.28395256 C  1.579867746
#> 79  -1.22123896 C  0.467376475
#> 80   0.27983296 C -0.599707072
#> 87  -0.88987647 C -1.098846392
#> 94  -0.89991063 C  0.310580823
#> 97   0.64964639 C  2.287231957
#> 100  0.53782163 C -0.675312161
#> 
#> [[2]]
#>              x y           z
#> 6  -0.42559900 C -0.67740184
#> 15 -0.86418063 A -0.14286970
#> 24  0.84916392 B  0.48086816
#> 31  0.18247646 B -0.67251918
#> 32 -0.44463685 A -0.27591741
#> 33  1.01665922 B  0.55100766
#> 34 -0.62688070 C  1.35060906
#> 38  0.65282819 A  0.41278476
#> 39  1.56194860 C -0.37909262
#> 48  1.21528847 C -0.15365442
#> 51 -2.75384847 A -0.20903429
#> 52  0.02274427 C  0.94400691
#> 53 -0.18781091 C -0.08139048
#> 57  1.00642906 C -1.60987109
#> 59  1.85553963 B  0.13146406
#> 67 -0.16196940 B -1.23605399
#> 70 -0.09029827 A  1.60662974
#> 71 -1.67358874 A -0.58148570
#> 75 -0.33737925 A  0.20849215
#> 81 -1.08577413 B -1.16315118
#> 82 -0.30454224 B -0.11324647
#> 83 -0.67115390 A  0.87598787
#> 85 -1.45293993 B  0.37379136
#> 89  0.98413276 A -0.86924041
#> 90  1.89543291 C -0.06089941
#> 93  0.24213645 A  1.52620410
#> 95 -1.35078331 C  0.78975964
#>