Clean and wrangle data

This function creates a data frame by grouping training data by predictor and strata_variables & adds a group_labels column

wrangle_training_data(training_data, predictor, strata_variable, group_labels)

Arguments

training_data

a dataframe for training data

predictor

a string representing the variable to be used as the predictor for the algorithm

strata_variable

a string representing the variable that will be set as the strata in initial_split

group_labels

list of strings representing the group names

Value

cleaned and wrangled training data grouped by predictor and strata_variable & add a label column

Examples

training_data <- data.frame(
  predictor = c("yes", "no", "yes", "no"),
  strata_variable = c("yes", "yes", "no", "no"),
  outcome = c(1, 1, 0, 1)
  )

 predictor <- "predictor"
 strata_variable <- "strata_variable"
 group_labels <- c("Group 1")

 wrangle_training_data(
  training_data, strata_variable, predictor,
  group_labels)
#> # A tibble: 4 × 4
#> # Groups:   strata_variable [2]
#>   strata_variable predictor     n label  
#>   <chr>           <chr>     <int> <chr>  
#> 1 no              no            1 Group 1
#> 2 no              yes           1 Group 1
#> 3 yes             no            1 Group 1
#> 4 yes             yes           1 Group 1