Split the dataset into training and testing sets based on a strata variable

Splits the data into testing/training and ensures variables are appropriate for classification algorithm

split_dataset(data_set, strata_variable, predictor)

Arguments

data_set

the data set to be used for the algorithm

strata_variable

a string representing the variable that will be set as the strata in initial_split

predictor

a string representing the variable to be used as the predictor for the algorithm

Value

a list containing the training and the testing data sets

Examples

data_set <- data.frame(
x = rnorm(100),
y = sample(c("A", "B", "C"), 100, replace = TRUE),
z = rnorm(100)
)
strata_variable <- "y"
predictor <- "x"

split_dataset(data_set, strata_variable, predictor)
#> [[1]]
#>               x y           z
#> 3   -1.58073072 A  1.35227149
#> 4   -0.82682433 A -0.03450955
#> 5   -1.79860207 A  0.22461668
#> 16  -0.86536170 A -0.43868990
#> 18   0.03329299 A -0.43079947
#> 20   0.55385257 A  0.86815550
#> 26  -1.23045212 A -1.08593855
#> 29  -1.20554347 A -2.26999366
#> 32   0.39421775 A  1.26684802
#> 38   0.89088650 A -0.50930499
#> 41  -0.38224291 A  0.82666785
#> 48   0.37627855 A  0.01677529
#> 50   0.65461936 A  1.46323103
#> 56   0.42428235 A  0.71008525
#> 57   0.50001433 A -0.03865365
#> 60   0.40086625 A  0.05958332
#> 65  -1.35678035 A  0.71021027
#> 70   0.67684385 A  0.58796338
#> 81   1.08950148 A  1.69879913
#> 82   0.10883692 A  0.06982855
#> 83  -0.87879798 A -0.44909137
#> 86  -0.64251789 A -1.62316103
#> 87  -0.40474214 A -0.12288366
#> 91   0.16024782 A -0.49087476
#> 96  -0.76270560 A -0.20260139
#> 97   0.14910751 A  1.35275551
#> 100  1.19221000 A  0.45454284
#> 12   0.17639220 B  0.92383633
#> 13   2.01901550 B -0.34536374
#> 17   0.02424781 B  1.12349939
#> 19  -0.62935138 B  0.72560626
#> 21  -0.15797970 B  1.04692796
#> 24  -0.36435855 B  1.50769934
#> 31  -0.24148605 B -1.78832863
#> 34   1.78187068 B  1.37384383
#> 35   0.68447196 B  0.23137665
#> 39   0.44427634 B -1.66917572
#> 42  -0.14631004 B  0.65834174
#> 53  -0.36419451 B -0.58291433
#> 54  -1.13087234 B -2.28363608
#> 63   0.58323727 B  0.06828652
#> 64   0.56032050 B  1.53246512
#> 68   0.55996878 B -1.33502461
#> 76   0.02438773 B -1.29083497
#> 77  -1.67300149 B  1.69296954
#> 79   1.49674554 B -0.18829650
#> 80   2.08459136 B -1.29972440
#> 85   0.41136848 B -1.99594712
#> 90   0.12838162 B -1.15539082
#> 2   -0.58787487 C  1.59088882
#> 7   -0.20600031 C -0.58600698
#> 8   -1.95156274 C -0.51584690
#> 14  -0.49896774 C -1.16846339
#> 15  -0.65652685 C -3.19032912
#> 22   1.46678587 C -1.12825523
#> 23   1.92197059 C  0.53434625
#> 25   0.53599415 C -0.84590640
#> 27  -2.31252700 C  0.25941128
#> 28  -0.35953511 C -3.39566048
#> 30   0.19666519 C  0.87755874
#> 33  -1.45858859 C  0.44066776
#> 37  -0.81813113 C -0.07567691
#> 47  -0.07375197 C  0.42357358
#> 49  -1.03183476 C  0.16617874
#> 58  -1.55850456 C -0.08450586
#> 62  -0.93328155 C  0.60674874
#> 72   0.54332757 C  0.04554945
#> 73   0.21936382 C  0.31657944
#> 74  -0.20706716 C  1.90424416
#> 78  -0.45630801 C -0.33501574
#> 84   1.29307365 C -2.35061102
#> 89   0.51872056 C  1.35897408
#> 92  -1.19413252 C  0.06696201
#> 99   0.03205666 C  0.03831723
#> 
#> [[2]]
#>              x y           z
#> 1   0.18008879 B  0.45250325
#> 6   0.88208150 A -0.29485464
#> 9   1.82124437 A  1.05486083
#> 10  0.55001897 C  0.59335450
#> 11 -0.08037915 A -0.69186087
#> 36 -0.54797807 C -0.54090922
#> 40 -0.76301682 C  0.30257438
#> 43 -0.61116088 A -0.22345407
#> 44  1.55185925 C  1.42363901
#> 45  1.05221311 B  0.65333274
#> 46 -0.02355922 B -0.10436199
#> 51 -0.49490927 B  1.79011939
#> 52  1.37370090 B  0.08365502
#> 55  0.24814246 B  0.06811675
#> 59 -0.64358125 A -0.32489415
#> 61 -0.90339381 C  1.11101304
#> 66 -0.52412670 C -0.33758990
#> 67  0.24914151 A -0.84604171
#> 69 -0.84954784 B -1.76665106
#> 71  0.91905347 A -1.34785736
#> 75 -1.25378543 C  1.62482582
#> 88 -0.03872443 A  1.53053620
#> 93 -0.29367746 B -0.39714447
#> 94  2.88243835 C  1.47533817
#> 95 -0.75681277 C  1.02776975
#> 98  0.20703212 A  2.26518787
#>