Split the dataset into training and testing sets based on a strata variable

Splits the data into testing/training and ensures variables are appropriate for classification algorithm

split_dataset(data_set, strata_variable, predictor)

Arguments

data_set

the data set to be used for the algorithm

strata_variable

a string representing the variable that will be set as the strata in initial_split

predictor

a string representing the variable to be used as the predictor for the algorithm

Value

a list containing the training and the testing data sets

Examples

data_set <- data.frame(
x = rnorm(100),
y = sample(c("A", "B", "C"), 100, replace = TRUE),
z = rnorm(100)
)
strata_variable <- "y"
predictor <- "x"

split_dataset(data_set, strata_variable, predictor)
#> [[1]]
#>              x y           z
#> 1  -1.66938491 A -0.26395506
#> 4   1.03158050 A -0.63846332
#> 6   0.29030909 A  0.77776308
#> 7  -0.54497523 A -0.27383959
#> 15  0.61029207 A -0.23925200
#> 17  1.43351943 A  0.77654931
#> 29 -0.33478580 A -0.25947763
#> 39  0.80651561 A -1.02454800
#> 42  0.56559545 A  0.61379683
#> 45 -1.38688851 A  0.11909690
#> 48  0.27561160 A -0.24820757
#> 56 -1.12027472 A -1.28169959
#> 67 -0.95449362 A -2.04892229
#> 71  1.86427484 A -0.44184890
#> 73  0.75357919 A -0.39263881
#> 75  0.23912777 A  1.21536518
#> 80  1.68939881 A -1.03586853
#> 82  0.06111287 A -1.15350270
#> 86 -0.18090644 A -0.28268129
#> 87 -0.89261894 A -0.84141139
#> 91  0.47728588 A -2.10606724
#> 93 -1.09287013 A -1.39662969
#> 98  0.95444489 A -0.20249050
#> 99  1.10619345 A  0.74470594
#> 2   3.42338504 B  1.54150969
#> 5  -0.41777194 B -0.22467693
#> 11  0.32029487 B  0.42351072
#> 12  0.71162403 B -0.19013474
#> 13 -0.71293833 B -0.26846196
#> 16 -1.14547165 B -1.66064701
#> 18 -1.29656906 B -1.22425840
#> 19 -0.48041490 B  0.33558829
#> 20  0.84464125 B  1.31878366
#> 25  0.05392105 B  0.16200895
#> 28 -1.25189717 B  0.29812732
#> 33 -1.47456113 B  0.48924551
#> 35  0.95318663 B  1.92066824
#> 40  0.50993798 B  0.16981217
#> 41  0.98095177 B -0.65132896
#> 44 -0.48748273 B  0.90640777
#> 49  0.38832404 B  1.07760676
#> 50  0.76362077 B -0.21699787
#> 52  0.51273913 B  0.54532162
#> 53 -0.20603766 B -0.20193002
#> 55  0.90633944 B -0.32882995
#> 59 -2.08561009 B  2.37045040
#> 60  0.47750086 B -1.89981682
#> 61 -1.29289962 B  0.23193489
#> 68  1.91276172 B -0.66101023
#> 69  2.60874926 B  0.70015454
#> 79 -0.22610235 B  1.69816823
#> 8  -0.43377066 C  0.92996552
#> 23  1.38112139 C  1.08183989
#> 26  1.05934167 C -0.29355257
#> 30 -2.50291151 C  1.38090699
#> 32  2.48759257 C -0.62560645
#> 43 -0.19571071 C -0.32057904
#> 51  0.36916680 C  0.31625908
#> 62 -2.36624172 C  0.29743614
#> 64  0.67351206 C  0.59960048
#> 66  1.01173063 C  1.09453835
#> 72  0.03706415 C -0.48625347
#> 74  1.34884282 C  0.29929914
#> 76  1.47644027 C -0.08760222
#> 77  0.35085108 C  0.29431151
#> 78 -0.98792517 C -0.60772060
#> 81  1.09129410 C  1.26302396
#> 84 -0.32669810 C  0.60890046
#> 88  1.13704875 C  1.84925261
#> 89 -1.15063849 C  1.73306482
#> 90  1.64622229 C  0.54656206
#> 94 -0.08350292 C -0.44475990
#> 95  0.84787640 C  2.12167276
#> 96 -0.65525571 C  0.91843013
#> 97  1.95466951 C  0.50363374
#> 
#> [[2]]
#>              x y           z
#> 3   -1.4582740 A -0.25444498
#> 9   -1.3713003 B -2.45621146
#> 10   0.4351102 B  0.40659572
#> 14  -1.3402739 B  2.10837708
#> 21   0.2097130 B -1.72457172
#> 22  -0.2466657 C  0.38885042
#> 24   0.1587692 C  2.34378153
#> 27  -2.1011845 A  0.03934379
#> 31   0.0704264 C -1.85568647
#> 34  -0.3702807 C -1.74765889
#> 36  -1.1728621 A -0.93607632
#> 37   2.6912379 B  0.96280617
#> 38  -0.4342503 A  1.27997012
#> 46  -0.5859983 A -0.59023012
#> 47  -0.9307724 C -1.26611126
#> 54  -0.8090778 C  0.03633203
#> 57   0.3939441 C  2.03927938
#> 58  -0.3254799 A  0.79805438
#> 63  -1.1967934 A  1.33546402
#> 65  -0.6007974 B  0.43219596
#> 70   2.2906132 B -1.59920951
#> 83  -0.3114794 C  0.02253587
#> 85   0.4172951 B  2.09387479
#> 92   0.2459562 B  0.53322402
#> 100 -0.3557065 A  1.44404450
#>