Split the dataset into training and testing sets based on a strata variable

Splits the data into testing/training and ensures variables are appropriate for classification algorithm

split_dataset(data_set, strata_variable, predictor)

Arguments

data_set

the data set to be used for the algorithm

strata_variable

a string representing the variable that will be set as the strata in initial_split

predictor

a string representing the variable to be used as the predictor for the algorithm

Value

a list containing the training and the testing data sets

Examples

data_set <- data.frame(
x = rnorm(100),
y = sample(c("A", "B", "C"), 100, replace = TRUE),
z = rnorm(100)
)
strata_variable <- "y"
predictor <- "x"

split_dataset(data_set, strata_variable, predictor)
#> [[1]]
#>               x y           z
#> 8  -0.806698691 A -0.13999195
#> 12 -0.169880457 A  0.30884746
#> 18  0.376947960 A -0.27326665
#> 29  1.601332775 A  0.75608385
#> 34 -1.672262865 A -0.89123632
#> 38 -0.259743903 A -1.37743259
#> 40 -0.634904567 A -1.19627780
#> 42  1.564465695 A  1.04620781
#> 44 -0.878075634 A  0.18545658
#> 47  0.214682596 A  0.01261102
#> 48  1.286600366 A -0.19182506
#> 54  0.068613467 A  0.16629641
#> 62 -0.864037124 A -1.90964936
#> 67  1.316929129 A  0.21871741
#> 68 -0.088260145 A  0.11631305
#> 75 -1.308546553 A  1.29290065
#> 76 -1.474930058 A -1.53533491
#> 77 -0.115281901 A  0.91203676
#> 79  0.941430225 A  0.21307753
#> 83  0.079415832 A -0.75900023
#> 85 -0.409772504 A  0.71864549
#> 92 -1.040574097 A -0.15884264
#> 95  0.702775091 A -0.12375319
#> 99  1.378098721 A  0.64100085
#> 7  -0.968045896 B -1.71707536
#> 11  2.562096726 B -0.62033684
#> 16 -1.355804474 B  0.37868183
#> 21  0.257331999 B  0.07778466
#> 24 -0.940954431 B  0.19329585
#> 25 -1.028183964 B  0.08550745
#> 30 -0.073979883 B -0.24128825
#> 32 -0.609426184 B -1.65395383
#> 33  0.971230375 B -1.99549604
#> 36  0.892481744 B -1.33228437
#> 37 -1.632646375 B -0.86077012
#> 53  0.974406241 B -0.53112618
#> 55 -0.593910419 B -0.81800245
#> 59  1.085946931 B  1.14442663
#> 63  0.425103407 B -0.70990417
#> 71  1.917793455 B  1.28033537
#> 82 -1.891249152 B  0.63539723
#> 90 -0.072108530 B -2.29794737
#> 91 -1.282117814 B  2.36807623
#> 93  0.338686994 B -1.38264610
#> 97  2.288848869 B -0.08221029
#> 2  -0.684658080 C  1.84451974
#> 6   0.640663015 C  0.99644374
#> 13  1.859923720 C  1.51685498
#> 15 -1.543493033 C -0.26393401
#> 19 -0.028113972 C -0.01490980
#> 20 -0.071110918 C -1.06997090
#> 26  0.671868644 C  0.15739189
#> 31  0.041353344 C -1.48911494
#> 39 -0.102590236 C  0.15774704
#> 41  0.553199638 C  1.08101468
#> 43 -1.267553564 C  0.31978085
#> 46 -1.999161707 C -0.53619715
#> 51  1.822738126 C  2.09131745
#> 52 -0.792276309 C  0.87608006
#> 56 -0.293279676 C -0.81985392
#> 58  0.643178799 C -1.10262512
#> 60  0.654022402 C  1.03188923
#> 61 -1.855417684 C  0.27606458
#> 64  0.177785311 C  0.06248528
#> 69 -0.151965979 C  3.13072808
#> 78  0.008854518 C -2.69328544
#> 80 -0.474829217 C -0.49755980
#> 81  0.643098924 C  0.38853970
#> 84  1.504494115 C  0.34544278
#> 87 -1.824568643 C -1.67372382
#> 88 -0.262178184 C  2.79204676
#> 89  0.564166632 C  0.24590603
#> 96 -1.589936954 C  1.04594671
#> 
#> [[2]]
#>               x y           z
#> 1    0.03567676 B -1.94839098
#> 3   -1.11404506 C -0.34888145
#> 4   -0.31939241 C  0.17932007
#> 5    2.35921650 C  0.88893022
#> 9    0.02854588 C  0.24894294
#> 10  -0.61967367 B  0.13929904
#> 14  -0.52912607 B  0.13577359
#> 17  -0.34925476 A  2.72480806
#> 22   0.06557126 B -0.17068149
#> 23   2.02532503 A  0.54800852
#> 27   1.30016951 C  1.35937284
#> 28  -1.13661946 A -0.07203580
#> 35   0.88677353 C  0.77565295
#> 45   0.30858576 B  1.75175847
#> 49  -0.16434448 A -0.75382071
#> 50  -0.76990301 A -1.11704486
#> 57  -1.49071599 C  0.79843418
#> 65  -0.64482467 B -0.33017804
#> 66   1.76989669 C -1.39952809
#> 70   0.61085346 A -0.04300988
#> 72  -2.48034341 A -0.18573802
#> 73   0.35694953 B  2.22196004
#> 74  -0.47359934 B  1.47193308
#> 86   1.32658644 A -0.03811867
#> 94  -1.07867164 C -0.92507760
#> 98  -0.41257050 C -0.41002205
#> 100  0.42465781 A -0.71787302
#>