Split the dataset into training and testing sets based on a strata variable

Splits the data into testing/training and ensures variables are appropriate for classification algorithm

split_dataset(data_set, strata_variable, predictor)

Arguments

data_set

the data set to be used for the algorithm

strata_variable

a string representing the variable that will be set as the strata in initial_split

predictor

a string representing the variable to be used as the predictor for the algorithm

Value

a list containing the training and the testing data sets

Examples

data_set <- data.frame(
x = rnorm(100),
y = sample(c("A", "B", "C"), 100, replace = TRUE),
z = rnorm(100)
)
strata_variable <- "y"
predictor <- "x"

split_dataset(data_set, strata_variable, predictor)
#> [[1]]
#>                x y           z
#> 5   0.0144914091 A -0.23381270
#> 9   0.0777022657 A -1.06837864
#> 18 -0.1823856378 A -0.50316403
#> 20  1.1213019612 A  0.30833156
#> 25  1.8702522459 A -0.04898191
#> 35  0.2349263360 A  0.95778009
#> 46  1.5719209314 A -1.54713022
#> 48  0.6020257654 A  0.25570497
#> 73 -0.0003781104 A  0.01661547
#> 84  1.2862464920 A -0.47293448
#> 88  0.0857140077 A -1.78203250
#> 92  0.1524178935 A -0.81481728
#> 94 -1.0682431235 A -1.25882186
#> 95  0.6956977509 A  1.47471237
#> 99  1.6338151823 A  0.52886031
#> 2   0.6359313887 B -0.32881860
#> 3  -0.5482257767 B -0.63438069
#> 7   0.5691762489 B  0.02807657
#> 10 -0.3648318736 B -0.65857678
#> 11  0.4211225728 B  1.42364527
#> 13 -1.0965314397 B -0.54812635
#> 22  0.1613389755 B  2.40419415
#> 24  0.9375527839 B -0.04429813
#> 28 -1.1789985595 B  0.52741458
#> 32 -0.7447051261 B  0.15895698
#> 38  1.6712615983 B -0.02015942
#> 39  1.4053975300 B -0.04902439
#> 51 -0.4580618222 B -0.54966285
#> 52  1.8594531578 B -0.02035061
#> 53  1.9583818300 B -0.70899992
#> 59  0.0879326624 B  0.46285506
#> 66 -1.4710755273 B  1.90158545
#> 67 -0.1072484756 B -1.01369628
#> 68  0.8498721050 B  0.01709082
#> 72 -0.5222977551 B  1.05574592
#> 77  0.0105890964 B -0.04545295
#> 80  1.1396826358 B  0.86720882
#> 85 -0.4583230647 B -0.78036532
#> 91  0.2468899114 B -0.67881520
#> 93  1.6284678848 B  1.68215093
#> 98 -1.6836938353 B -1.19243855
#> 6  -0.9074494599 C -0.23603778
#> 8  -1.6797037114 C -0.28883193
#> 12 -0.0050738545 C  1.02761287
#> 14  0.4661646541 C -1.49443466
#> 15  0.0185804485 C -0.82882899
#> 21  1.5773336383 C  1.02576822
#> 27 -0.0530923312 C -1.22102553
#> 29  1.9192514090 C  0.75550585
#> 30  0.3727499965 C -0.62409056
#> 31 -0.1767734631 C  1.68346922
#> 34  0.6509279203 C  0.97202620
#> 36  0.1152846967 C -0.99814061
#> 42  0.5790336224 C -2.27285841
#> 45 -0.7429726089 C  0.40496151
#> 47  0.8025451159 C -0.27672030
#> 49 -0.8983013749 C  0.15730896
#> 55 -0.1585854680 C  0.38804385
#> 57  0.5815694705 C  0.17223341
#> 58 -0.2925204695 C -0.12009157
#> 60 -1.5540520769 C  0.73181255
#> 61 -1.2415637710 C  0.03893157
#> 62 -0.0557317714 C  0.44166582
#> 63  0.0269175679 C  0.18787751
#> 65 -0.4078400764 C  0.28253450
#> 69 -0.0711435154 C -0.28905973
#> 75 -0.4015424035 C -1.76043169
#> 79  1.0504725991 C -0.27351290
#> 82 -0.5377526887 C  1.79572123
#> 83  0.4863664640 C  0.89501544
#> 87 -0.6439008375 C  1.08416724
#> 89  1.2663870542 C  1.53371499
#> 90  1.2036887129 C  0.43157812
#> 97  0.6568409955 C -1.07544719
#> 
#> [[2]]
#>               x y          z
#> 1    1.52900303 B  0.2698919
#> 4    0.02360995 C  1.2298968
#> 16   0.57195037 B  0.5860680
#> 17  -0.32448616 C  0.3103465
#> 19   0.76722478 C  0.5741413
#> 23  -0.81503006 A  0.8857195
#> 26   1.14626661 B -1.6671171
#> 33  -0.66448197 B -0.9922000
#> 37  -0.35898417 A  0.4649553
#> 40  -0.65780801 C  2.2847280
#> 41   1.01432874 C  0.2108871
#> 43  -0.59422476 B -1.4346514
#> 44   1.23889172 C  1.4910548
#> 50   0.80356158 C  1.0577567
#> 54   0.29258426 B  1.0800102
#> 56  -0.89314684 B -0.4369813
#> 64   0.08897939 C -1.8904177
#> 70  -0.93920561 B -0.2161871
#> 71  -0.98239874 C  0.6271887
#> 74   0.93706856 B  0.5393876
#> 76   0.37811144 C  1.4636922
#> 78   0.61024002 A -0.4323860
#> 81  -0.41445221 A  2.0684215
#> 86   1.56830801 A -0.1269690
#> 96   0.59860835 C  1.4715765
#> 100 -1.53205081 A -1.1951391
#>